Added validation check in pipeline install command
